refactor(packaging): view
